Abstract UseOrgUnits component
